refactor: move archive logic to its own module
