+++ src/fs/mod.rs
@@ -1,1 +1,3 @@
-pub mod directories;
+pub mod archive;
+pub mod directory;
+pub mod symlink;
